requirements.txt
streamlit
openai
pypdf
fpdf2
